Initial reref

+++ vamtb/ref/__init__.py
@@ -15,14 +15,14 @@ def vmb_vmi(refi):
         if fn.endswith(".vmi"):
             fnb = fn[0:-1] + "b"
             if fnb not in refi:
-                warn(f"We found a reference for {fn} but not its counterpar {fnb}")
+                warn(f"We found a reference for {fn} but not its counterpart {fnb}")
                 continue
             refo[fn] = refi[fn]
             refo[fnb] = refi[fnb]
         elif fn.endswith(".vmb"):
             fni = fn[0:-1] + "i"
             if fni not in refi:
-                warn(f"We found a reference for {fn} but not its counterpar {fni}")
+                warn(f"We found a reference for {fn} but not its counterpart {fni}")
                 continue
             refo[fn] = refi[fn]
             refo[fni] = refi[fni]
@@ -33,7 +33,7 @@ def vmb_vmi(refi):
 def modify_files(var, newref):
     tdir = var.tmpDir
     for globf in search_files_indir(tdir, "*"):
-        if globf.name == "meta.json":
+        if globf.name == "meta.json" or globf.suffix in (".vmi", ".vam", ".vab", ".assetbundle", ".tif", ".jpg", ".png", ".dll"):
             continue
         try:
             file = FileName(globf)
@@ -41,13 +41,15 @@ def modify_files(var, newref):
         except:
             #Not a json file
             continue
+        info(f"> Searching for pattern in {globf}")
         for nr in newref:
+#            info(f">> Applying reref for {nr}")
             with open(globf, 'r+') as f:
                 fs = f.read()
                 rep = f"{newref[nr]['newvar']}:/{newref[nr]['newfile']}"
                 replace_string = fs.replace(f"SELF:/{nr}", rep).replace(f"/{nr}", rep)
                 if replace_string != fs:
-                    debug(f"Rerefing {globf} for {nr} to point to {rep}")
+                    info(f">>> Rerefing {globf} for {nr} to point to {rep}")
                     f.seek(0)
                     f.write(replace_string)
 
@@ -70,9 +72,13 @@ def modify_meta(var:Var, newref):
         f.write(prettyjson(meta))
     logging.debug("Modified meta")
 
-def reref_var(var:Var):
+def reref_var(var:Var, dryrun=True):
     new_ref = {}
     var_already_as_ref = []
+    info(f"Searching for dups in {var}")
+    if Dbs.get_ref(var.var) == "YES":
+        warn(f"Asked to reref {var} but it is a reference var, not doing anything.")
+        return
     for file in Dbs.get_files(var.var):
         debug(f"Checking if {file} has a reference candidate")
         choice = 0
@@ -94,7 +100,15 @@ def reref_var(var:Var):
             new_ref[file]['newvar'] = ref[0]
             new_ref[file]['newfile'] = ref[1]
     new_ref = vmb_vmi(new_ref)
-    modify_meta(var, new_ref)
-    modify_files(var, new_ref)
-    delete_files(var, new_ref)
-    pass
+    if not new_ref:
+        info("Found nothing to reref")
+        return
+    else:
+        info(f"Found these files as duplicates:{','.join(list(new_ref))}")
+    if dryrun:
+        info("Asked for dryrun, stopping here")
+    else:
+        modify_meta(var, new_ref)
+        modify_files(var, new_ref)
+        delete_files(var, new_ref)
+        zipdir(var.tmpDir, var.file)
